backend/requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
Pillow==10.4.0
